Move policy mock for external use
