+++ pkg/attest/verify_test.go
@@ -8,7 +8,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/docker/attest/internal/test"
 	"github.com/docker/attest/pkg/attestation"
 	"github.com/docker/attest/pkg/oci"
 	"github.com/docker/attest/pkg/policy"
@@ -43,9 +42,9 @@ func TestVerifyAttestations(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 
-			mockPE := test.MockPolicyEvaluator{
+			mockPE := policy.MockPolicyEvaluator{
 				EvaluateFunc: func(ctx context.Context, resolver oci.AttestationResolver, pfs []*policy.PolicyFile, input *policy.PolicyInput) (*rego.ResultSet, error) {
-					return test.AllowedResult(), tc.policyEvaluationError
+					return policy.AllowedResult(), tc.policyEvaluationError
 				},
 			}
